+++ api/blueprints/drbEdition.py
@@ -21,12 +21,15 @@ def editionFetch(editionID):
         or current_app.config['READER_VERSION']
 
     edition = dbClient.fetchSingleEdition(editionID)
+
+    editionWorkTitle = edition.work.title
+    editionWorkAuthors = workAuthorInfo(edition)
             
     if edition:
         statusCode = 200
         records = dbClient.fetchRecordsByUUID(edition.dcdw_uuids)
         responseBody = APIUtils.formatEditionOutput(
-            edition, records=records, showAll=showAll, reader=readerVersion
+            edition, editionWorkTitle, editionWorkAuthors, records=records, showAll=showAll, reader=readerVersion
         )
     else:
         statusCode = 404
@@ -43,3 +46,10 @@ def editionFetch(editionID):
     return APIUtils.formatResponseObject(
         statusCode, 'singleEdition', responseBody
     )
+
+# Return a list of authors to formatEdition method
+def workAuthorInfo(edition):
+    workAuthors = []
+    for i in edition.work.authors:
+        workAuthors.append(i['name'])
+    return workAuthors